version 0.2.9 - EstimateExpireDate use Long instead of Int

+++ dataclass/src/main/kotlin/NoExpDB.kt
@@ -10,8 +10,8 @@ data class NoExpDB(
     val home: Map<String, ProductDao>? = null,
     val lastQr: String? = null,
     val category: Map<String, CategoryDao>? = null,
-    val expireDateByCategory: Map<String, Int>? = null,
-    val expireDateByBarcode: Map<String, Int>? = null
+    val expireDateByCategory: Map<String, Long>? = null,
+    val expireDateByBarcode: Map<String, Long>? = null
 ) : NoExpDBModel() {
     override fun toJson() = json.stringify(serializer(), this)
